+++ components/DateTimePicker/date-time-picker.vue
@@ -19,6 +19,7 @@
                     }"
                     @focus="syncDateTime(inputDateString)"
                     @keyup.enter="setDate(inputDateString)"
+                    @blur="setDate(inputDateString)"
             >
                 <template
                         v-if="iconPosition==='prefix'"