+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 更新日志
 
 ## v1.0.65
+### Optimization
+  * table
+    * 优化filter筛选交互
 ### Bug fixes
   * table
     * fix 无hover行状态